% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{arbuthnot_tbl}
\alias{arbuthnot_tbl}
\title{Male and female births in London}
\format{
A tibble with 82 observations and 3 variables:
\describe{
  \item{year}{Year of record (integer).}
  \item{boys}{Number of boys baptized (integer).}
  \item{girls}{Number of girls baptized (integer).}
}
}
\source{
Arbuthnot, J. (1710). *An argument for Divine Providence, taken from the constant regularity observed in the births of both sexes*.
}
\usage{
data(arbuthnot_tbl)
}
\description{
The `arbuthnot_tbl` dataset contains historical baptism records collected by John Arbuthnot.
It includes data on the number of boys and girls baptized each year from 1629 to 1710 in London.
This dataset was historically significant in one of the earliest applications of statistical
analysis to human populations. The dataset name has been suffixed with `_tbl` to indicate
that it is a tibble, ensuring clarity within the `sampledatasets` package.
}
