% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{cards_tbl}
\alias{cards_tbl}
\title{Standard Deck of Playing Cards}
\format{
A tibble with 52 observations and 4 variables:
\describe{
  \item{value}{Card value (factor with 13 levels: "2" to "10", "Jack", "Queen", "King", "Ace").}
  \item{color}{Card color (factor with 2 levels: "black", "red").}
  \item{suit}{Card suit (factor with 4 levels: "Clubs", "Diamonds", "Hearts", "Spades").}
  \item{face}{Logical indicator (TRUE if the card is a face card, FALSE otherwise).}
}
}
\source{
Constructed dataset representing a standard deck of cards.
}
\usage{
data(cards_tbl)
}
\description{
The `cards_tbl` dataset represents a standard deck of 52 playing cards. It includes information
about the value, color, suit, and whether the card is a face card. The dataset name has been
suffixed with `_tbl` to indicate that it is a tibble, ensuring clarity within the `sampledatasets` package.
}
