% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{cars_df}
\alias{cars_df}
\title{Speed and Stopping Distances of Cars}
\format{
A data frame with 50 observations and 2 variables:
\describe{
  \item{speed}{Speed of the car (mph) (numeric).}
  \item{dist}{Stopping distance (ft) (numeric).}
}
}
\source{
Data from the 1920s. Originally included in the *datasets* package in R.
}
\usage{
data(cars_df)
}
\description{
The `cars_df` dataset contains data on the speed of cars and the corresponding distances required
to stop. This dataset is often used in statistical analysis and regression modeling. The dataset
name has been suffixed with `_df` to indicate that it is a data frame, ensuring clarity within
the `sampledatasets` package.
}
