% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{mtcars_df}
\alias{mtcars_df}
\title{Motor Trend Car Road Tests}
\format{
A data frame with 32 observations and 11 variables:
\describe{
  \item{mpg}{Miles/(US) gallon (numeric).}
  \item{cyl}{Number of cylinders (numeric).}
  \item{disp}{Displacement (cu.in.) (numeric).}
  \item{hp}{Gross horsepower (numeric).}
  \item{drat}{Rear axle ratio (numeric).}
  \item{wt}{Weight (1000 lbs) (numeric).}
  \item{qsec}{1/4 mile time (numeric).}
  \item{vs}{Engine (0 = V-shaped, 1 = straight) (numeric).}
  \item{am}{Transmission (0 = automatic, 1 = manual) (numeric).}
  \item{gear}{Number of forward gears (numeric).}
  \item{carb}{Number of carburetors (numeric).}
}
}
\source{
Motor Trend, 1974.
}
\usage{
data(mtcars_df)
}
\description{
The `mtcars_df` dataset provides information extracted from the 1974 Motor Trend US magazine.
It comprises data on fuel consumption and 10 aspects of automobile design and performance
for 32 cars (1973–74 models). The dataset name has been suffixed with `_df` to identify
it as a data frame, making it easier to distinguish within the `sampledatasets` package.
}
