% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledatasets-package.R
\docType{package}
\name{usdatasets}
\alias{sampledatasets}
\alias{usdatasets}
\alias{sampledatasets-package}
\title{sampledatasets: A Collection of Sample Datasets}
\description{
{
This package provides a variety of datasets related to automotive performance and safety data to historical demographics and socioeconomic indicators, as well as recreational data.
}
}
\details{
sampledatasets: A Collection of Sample Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Collection of Sample Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/LatinR_2024_R_DEVELOPER}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Cáceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
